feat(default-theme): prop for columns count in footer navigation (#1904)
* Add prop for columns

Currently the number of columns is hard-coded, so that - when people use different slot content for this component - the number of columns does not match up. With this PR, people can simply pass in a prop from the parent component (or a wrapper component `src/components/organisms/SwFooterNavigation.vue`) to change the number.

* Remove ref

On the other hand, this does not need to be a `ref` anyway, because nothing is changing the variable on the fly. It is static, earlier as a hard-coded value, now as a prop.

* Props should not be redeclared

Export a new variable with same name as a prop from the `setup` is not going to work. It only needs to be declared in `props`.